Move the stringifier related classes to the `org.spine3.base` package.
